requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
argcomplete==1.12.3
argon2-cffi==21.1.0
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
blis==0.7.5
cached-property==1.5.2
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cymem==2.0.6
DAWG-Python==0.7.2
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
faiss-cpu==1.7.1.post2
Flask==2.0.2
frozenlist==1.2.0
gensim==4.1.2
h5py==3.5.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
joblib==1.1.0
jsonpickle==2.0.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
langcodes==3.3.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
murmurhash==1.0.6
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.5
notebook==6.4.5
numpy==1.21.4
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
preshed==3.0.6
prometheus-client==0.12.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
regex==2021.11.10
requests==2.26.0
rnnmorph==0.4.1
ru-core-news-md @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_md-3.2.0/ru_core_news_md-3.2.0-py3-none-any.whl
russian-tagsets==0.6
scikit-learn==1.0.1
scipy==1.7.2
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
terminado==0.12.1
testpath==0.5.0
thinc==8.0.13
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
yarl==1.7.2
zipp==3.6.0
